added repositories, blocs, cubits, pages + home page logic

+++ modules/features/lib/widgets/top_up_card.dart
@@ -3,8 +3,8 @@ import 'package:features/widgets/colors.dart';
 import 'package:flutter/material.dart';
 
 class TopUpCard extends StatelessWidget {
-  final int amount;
-  final Function(int amount) onPressed;
+  final double amount;
+  final Function(double amount) onPressed;
 
   const TopUpCard({
     super.key,
